Service Manager Separation
