+++ pkg/services/mysqldb/deprovision_all_in_one.go
@@ -1,0 +1,58 @@
+package mysqldb
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/Azure/open-service-broker-azure/pkg/service"
+)
+
+func (a *allInOneManager) GetDeprovisioner(
+	service.Plan,
+) (service.Deprovisioner, error) {
+	return service.NewDeprovisioner(
+		service.NewDeprovisioningStep("deleteARMDeployment", a.deleteARMDeployment),
+		service.NewDeprovisioningStep("deleteMySQLServer", a.deleteMySQLServer),
+	)
+}
+
+func (a *allInOneManager) deleteARMDeployment(
+	_ context.Context,
+	instance service.Instance,
+) (service.InstanceDetails, error) {
+	dt, ok := instance.Details.(*allInOneMysqlInstanceDetails)
+	if !ok {
+		return nil, fmt.Errorf(
+			"error casting instance.Details as *allInOneMysqlInstanceDetails",
+		)
+	}
+	if err := a.armDeployer.Delete(
+		dt.ARMDeploymentName,
+		instance.ResourceGroup,
+	); err != nil {
+		return nil, fmt.Errorf("error deleting ARM deployment: %s", err)
+	}
+	return dt, nil
+}
+
+func (a *allInOneManager) deleteMySQLServer(
+	_ context.Context,
+	instance service.Instance,
+) (service.InstanceDetails, error) {
+	dt, ok := instance.Details.(*allInOneMysqlInstanceDetails)
+	if !ok {
+		return nil, fmt.Errorf(
+			"error casting instance.Details as *allInOneMysqlInstanceDetails",
+		)
+	}
+	cancelCh := make(chan struct{})
+	_, errChan := a.serversClient.Delete(
+		instance.ResourceGroup,
+		dt.ServerName,
+		cancelCh,
+	)
+	if err := <-errChan; err != nil {
+		return nil, fmt.Errorf("error deleting mysql server: %s", err)
+	}
+	return dt, nil
+}
