+++ pkg/services/mysqldb/deprovision_db_only.go
@@ -1,0 +1,63 @@
+package mysqldb
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/Azure/open-service-broker-azure/pkg/service"
+)
+
+func (d *dbOnlyManager) GetDeprovisioner(
+	service.Plan,
+) (service.Deprovisioner, error) {
+	return service.NewDeprovisioner(
+		service.NewDeprovisioningStep("deleteARMDeployment", d.deleteARMDeployment),
+		service.NewDeprovisioningStep("deleteMySQLServer", d.deleteMySQLServer),
+	)
+}
+
+func (d *dbOnlyManager) deleteARMDeployment(
+	_ context.Context,
+	instance service.Instance,
+) (service.InstanceDetails, error) {
+	dt, ok := instance.Details.(*dbOnlyMysqlInstanceDetails)
+	if !ok {
+		return nil, fmt.Errorf(
+			"error casting instance.Details as *dbOnlyMysqlInstanceDetails",
+		)
+	}
+	if err := d.armDeployer.Delete(
+		dt.ARMDeploymentName,
+		instance.ResourceGroup,
+	); err != nil {
+		return nil, fmt.Errorf("error deleting ARM deployment: %s", err)
+	}
+	return dt, nil
+}
+func (d *dbOnlyManager) deleteMySQLServer(
+	_ context.Context,
+	instance service.Instance,
+) (service.InstanceDetails, error) {
+	pdt, ok := instance.Parent.Details.(*dbmsOnlyMysqlInstanceDetails)
+	if !ok {
+		return nil, fmt.Errorf(
+			"error casting instance.Details as *dbOnlyMysqlInstanceDetails",
+		)
+	}
+	dt, ok := instance.Details.(*dbOnlyMysqlInstanceDetails)
+	if !ok {
+		return nil, fmt.Errorf(
+			"error casting instance.Details as *dbOnlyMysqlInstanceDetails",
+		)
+	}
+	cancelCh := make(chan struct{})
+	_, errChan := d.serversClient.Delete(
+		instance.ResourceGroup,
+		pdt.ServerName,
+		cancelCh,
+	)
+	if err := <-errChan; err != nil {
+		return nil, fmt.Errorf("error deleting mysql server: %s", err)
+	}
+	return dt, nil
+}
