+++ pkg/services/mysqldb/provision_common.go
@@ -1,0 +1,74 @@
+package mysqldb
+
+import (
+	"bytes"
+	"errors"
+	"fmt"
+	"net"
+	"strings"
+
+	"github.com/Azure/open-service-broker-azure/pkg/service"
+	_ "github.com/go-sql-driver/mysql" // MySQL driver
+)
+
+func validateServerProvisionParameters(
+	provisioningParameters service.ProvisioningParameters,
+) error {
+	pp, ok := provisioningParameters.(*ServerProvisioningParameters)
+	if !ok {
+		return errors.New(
+			"error casting provisioningParameters " +
+				"as *mysql.ServerProvisioningParameters",
+		)
+	}
+	sslEnforcement := strings.ToLower(pp.SSLEnforcement)
+	if sslEnforcement != "" && sslEnforcement != enabled &&
+		sslEnforcement != disabled {
+		return service.NewValidationError(
+			"sslEnforcement",
+			fmt.Sprintf(`invalid option: "%s"`, pp.SSLEnforcement),
+		)
+	}
+	if pp.FirewallIPStart != "" || pp.FirewallIPEnd != "" {
+		if pp.FirewallIPStart == "" {
+			return service.NewValidationError(
+				"firewallStartIPAddress",
+				"must be set when firewallEndIPAddress is set",
+			)
+		}
+		if pp.FirewallIPEnd == "" {
+			return service.NewValidationError(
+				"firewallEndIPAddress",
+				"must be set when firewallStartIPAddress is set",
+			)
+		}
+	}
+	startIP := net.ParseIP(pp.FirewallIPStart)
+	if pp.FirewallIPStart != "" && startIP == nil {
+		return service.NewValidationError(
+			"firewallStartIPAddress",
+			fmt.Sprintf(`invalid value: "%s"`, pp.FirewallIPStart),
+		)
+	}
+	endIP := net.ParseIP(pp.FirewallIPEnd)
+	if pp.FirewallIPEnd != "" && endIP == nil {
+		return service.NewValidationError(
+			"firewallEndIPAddress",
+			fmt.Sprintf(`invalid value: "%s"`, pp.FirewallIPEnd),
+		)
+	}
+	//The net.IP.To4 method returns a 4 byte representation of an IPv4 address.
+	//Once converted,comparing two IP addresses can be done by using the
+	//bytes. Compare function. Per the ARM template documentation,
+	//startIP must be <= endIP.
+	startBytes := startIP.To4()
+	endBytes := endIP.To4()
+	if bytes.Compare(startBytes, endBytes) > 0 {
+		return service.NewValidationError(
+			"firewallEndIPAddress",
+			fmt.Sprintf(`invalid value: "%s". must be 
+				greater than or equal to firewallStartIPAddress`, pp.FirewallIPEnd),
+		)
+	}
+	return nil
+}
